Legg opp til å kunne bruke manuelt endra tekst som inndata generelt
