+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/etterlatte/maler/BrevDTO.kt
@@ -1,0 +1,5 @@
+package no.nav.pensjon.etterlatte.maler
+
+interface BrevDTO {
+    val innhold: List<Element>
+}
