+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/etterlatte/maler/ElementTilBrevbakerkonvertering.kt
@@ -1,19 +1,21 @@
 package no.nav.pensjon.etterlatte.maler
 
+import no.nav.pensjon.brev.template.Expression
 import no.nav.pensjon.brev.template.LangBokmalNynorskEnglish
 import no.nav.pensjon.brev.template.Language
 import no.nav.pensjon.brev.template.dsl.OutlineOnlyScope
 import no.nav.pensjon.brev.template.dsl.expression.equalTo
 import no.nav.pensjon.brev.template.dsl.textExpr
-import no.nav.pensjon.etterlatte.maler.ManueltBrevDTOSelectors.ElementSelectors.children
-import no.nav.pensjon.etterlatte.maler.ManueltBrevDTOSelectors.ElementSelectors.type
-import no.nav.pensjon.etterlatte.maler.ManueltBrevDTOSelectors.InnerElementSelectors.children
-import no.nav.pensjon.etterlatte.maler.ManueltBrevDTOSelectors.InnerElementSelectors.text
-import no.nav.pensjon.etterlatte.maler.ManueltBrevDTOSelectors.innhold
+import no.nav.pensjon.etterlatte.maler.ElementSelectors.children
+import no.nav.pensjon.etterlatte.maler.ElementSelectors.type
+import no.nav.pensjon.etterlatte.maler.InnerElementSelectors.children
+import no.nav.pensjon.etterlatte.maler.InnerElementSelectors.text
 
-fun OutlineOnlyScope<LangBokmalNynorskEnglish, ManueltBrevDTO>.konverterElementerTilBrevbakerformat() {
+fun <D : BrevDTO> OutlineOnlyScope<LangBokmalNynorskEnglish, D>.konverterElementerTilBrevbakerformat(
+    innhold: Expression<List<Element>>,
+) {
     forEach(innhold) { element ->
-        showIf(element.type.equalTo(ManueltBrevDTO.ElementType.HEADING_TWO)) {
+        showIf(element.type.equalTo(ElementType.HEADING_TWO)) {
             forEach(element.children) { inner ->
                 title1 {
                     ifNotNull(inner.text) {
@@ -25,7 +27,7 @@ fun OutlineOnlyScope<LangBokmalNynorskEnglish, ManueltBrevDTO>.konverterElemente
                     }
                 }
             }
-        }.orShowIf(element.type.equalTo(ManueltBrevDTO.ElementType.HEADING_THREE)) {
+        }.orShowIf(element.type.equalTo(ElementType.HEADING_THREE)) {
             forEach(element.children) { inner ->
                 title2 {
                     ifNotNull(inner.text) {
@@ -37,7 +39,7 @@ fun OutlineOnlyScope<LangBokmalNynorskEnglish, ManueltBrevDTO>.konverterElemente
                     }
                 }
             }
-        }.orShowIf(element.type.equalTo(ManueltBrevDTO.ElementType.PARAGRAPH)) {
+        }.orShowIf(element.type.equalTo(ElementType.PARAGRAPH)) {
             paragraph {
                 forEach(element.children) { inner ->
                     ifNotNull(inner.text) {
@@ -49,7 +51,7 @@ fun OutlineOnlyScope<LangBokmalNynorskEnglish, ManueltBrevDTO>.konverterElemente
                     }
                 }
             }
-        }.orShowIf(element.type.equalTo(ManueltBrevDTO.ElementType.BULLETED_LIST)) {
+        }.orShowIf(element.type.equalTo(ElementType.BULLETED_LIST)) {
             paragraph {
                 list {
                     forEach(element.children) { inner ->
