+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/etterlatte/maler/omstillingsstoenad/OMSInnvilgelseManuell.kt
@@ -9,6 +9,7 @@ import no.nav.pensjon.brevbaker.api.model.LetterMetadata
 import no.nav.pensjon.etterlatte.EtterlatteBrevKode
 import no.nav.pensjon.etterlatte.EtterlatteTemplate
 import no.nav.pensjon.etterlatte.maler.ManueltBrevDTO
+import no.nav.pensjon.etterlatte.maler.ManueltBrevDTOSelectors.innhold
 import no.nav.pensjon.etterlatte.maler.fraser.common.OMSFelles
 import no.nav.pensjon.etterlatte.maler.fraser.common.Vedtak
 import no.nav.pensjon.etterlatte.maler.konverterElementerTilBrevbakerformat
@@ -39,7 +40,7 @@ object OMSInnvilgelseManuell : EtterlatteTemplate<ManueltBrevDTO> {
         outline {
             includePhrase(Vedtak.Overskrift)
 
-            konverterElementerTilBrevbakerformat()
+            konverterElementerTilBrevbakerformat(innhold)
 
             includePhrase(OMSFelles.MeldFraOmEndringer)
             includePhrase(OMSFelles.DuHarRettTilAaKlage)
